Seguridad F

+++ src/main/java/com/dh/clinica/repository/impl/DomicilioRepository.java
@@ -3,5 +3,5 @@ package com.dh.clinica.repository.impl;
 import com.dh.clinica.model.Domicilio;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DomicilioRepository extends JpaRepository<Domicilio, Long> {
+public interface DomicilioRepository extends JpaRepository<Domicilio, Integer> {
 }
